Remove Routes
Removes all signs of Routes from the Fn server.

Fn Call have been changed to use fn_id over the route path. This has
led to many changes in the code base. Note, the Routes table does
remain, as existing migrations modifiy it, a new migration drops it,
but changing migrations is verboten, so it has to stay.