fn: removed hotLauncher in favor of requests handling same task
Revert to older launch strategy where each request makes
an individual decision on container launch as opposed to
hot launcher nanny go-routine per slot queue.

Given the existence of the evictor, this simplifies
this decision logic and allows us to be more aggressive
in launch decisions.

The new strategy is as follows:

1) Each request checks slot queue first to acquire
    a free slot. If not, checkLaunch() is called.

2) Main loop of checkLaunch() monitors two main sets of events
    at a time, for example:

    A) free slots + new container launch
    B) free slots + resource wait

    and it switches between these two sets depending on
    the state of the main loop. Resource tracker no longer
    returns CapacityFull and it no longer has non-blocking mode.
    If a resource is not returned in HotPoll interval,
    an eviction is attempted and the result of the eviction
    is propagated to callers if non-blocking mode is enabled.