Span to View converter
We have added open census span calls in interesting parts of the code,
but without creating a jaeger or zipkin cluster they are not going
anywhere.

Here we add a SpanConverter that converts our Spans into OpenCensus
distribution Views.

It contains an in memory map of the Measures for these Views,
recording values for the length of time reported to be spent in a
Span.

This list has a hard limit of 100, as there is a risk of a metric
explosion should an element of the system include some ID value in
it's Span names. Gin does this and those are discarded, but there
might be others that we'll have to deal with in the future when discovered.