Favour user fn response codes for triggers, with caveats
Currently, this is a little unclear and there is a problem with user
fns returning error responses which we can not differentiate from our
service errors.

This changes breaks the status selection logic out and adds some tests that spell out
how we want this fiddly bit to work. The offering here is: user fn
status returned, if set, but modify 5xx to 502. If no response code
from user fn, use one the service has deemed appropriate. Default to
200 if neither of them have commented.