fn: restructure Cookie/Wait api
As defined in Driver interface, PrepareCookie() can
include any preparation that is retriable. With this
change, in PrepareCookie() in addition to image pull
and container create, we also attach to the container
and start stats collection.

This simplifies the interface and keeps Cookie.Run()
lean, which merely starts task & container.