fn: service versus user origin errors
502/504 http status codes are reserved for user-code
related errors in Fn. Some examples of these are:

    * user-container did not respond in time
    * user-container did not initialize in time
    * FDK returned 502 or 504

This PR introduces the following changes:

* docker-pull timeout is now a 504 which classifies it as a
user-code related error.

* docker-pull errors which return 5xx are now using
502/504 which classifies these as user-code related errors.

* Only applicable to detached mode, a timeout on LB is
now a ErrServiceReservationFailure (500). In detached mode,
this is unlikely to make it back to a client and it is mostly
for documentation/metrics purposes.

* Client quit/cancel cases are now using http 499. This is
an unofficial http code used by Nginx. This error code is
unlikely to make it back to the client and it is mostly
for documentation/metrics purposes.

* For Triggers, user versus service status code changes:
        * service http code takes priority
        * if service http code is 2xx, only then we decap user-http-code
        * user-http-code can return 502 and 504, but other 5xx codes
          are scrubbed into a 502.