adds tests for Fn-Call-Id, and adds for timeouts
back when we streamed responses out from the container, we set this header
regardless of whether the task ran or not. now, we're setting this header at
the end since we're doing buffering, only if the task runs successfully atm.
we can change to the old policy pretty readily if we want to, up for
discussion. as this patch presents it, it's just adding a test that it exists
in the normal case and the timeout case, and adding the timeout case. however,
other relevant cases exist, like fdk protocol errors, and we can whitelist
those each or just cover all errors, even if the call was never executed, up
to us. when we add streaming, the plumbing goes back to the way it was where
we just add this at some point before the container writes any headers out of
it, and we can pick what level to do that if we want special behavior around
this header for whether the call executed or not.