remove unused images when total image size exceeds a limit (#1364)
* fn: introducing image cleaner cache

Introducing a new image cleaner background job in docker driver.
If enabled (with FN_IMAGE_CLEAN_MAX_SIZE set > 0 bytes),
the cleaner performs the following at startup:

*) list images to populate the image cache, once this completes, then
*) in a loop, wait for notifications from image cache if/when
   total cache size exceeds max. Periodically pick a LRU image
   and delete it in docker in order to meet the cache size constraint.

The cleaner is not tested and will likely fail if docker build environment is present since these
can leave intermediate and/or dangling images. Also the cleaner expects that no outside manual
action is performed on these images. For example, a force removal can leave an image untagged/dangling. Cleaner interaction with these environments/cases is not tested.

FN_IMAGE_CLEAN_EXEMPT_TAGS also allows to exempt various tags from
the image cleaner.

When hot containers are running, they allocate a Cookie. These cookies
set their images "in-use" if Inspect/Pull succeeds. These "in-use"
images are reference counted and moved to a busy-list internally
in the image cleaner. "in-use" images are not stored in the LRU cache
and therefore cleaner job will not try to delete them. During termination
of the Cookie objects, image reference count is decremented and once
it hits zero, only then that image becomes eligible for removal
in the LRU cache.

Occasionally, the image cleaner may make the wrong call and remove
an image that was just Pulled or Inspected by a new Cookie. Here
the behavior is optimistic and assumes that this will be rare.
If it occurs, then CreateContainer emits a 503 "server-too-busy"
error which can be retried on the same or a different runner.
Pathological cases in which requests trigger removal of the other
can be monitored by CreateContainer failure and RemoveImage call
counters.