adds user blame tag for trigger/invoke response metric (#1348)
* adds 'blame' tag to response api metrics

we are attempting to ascertain whether a status code is from a function or
from the platform for invoke and http triggers. this proposes that some errors
(models.APIError) are at the same level of fault to the user as a user
returning a given status code from their trigger or from the container,
examples of such errors are: docker 401, timeout 504, bad function response
502. the thinking being, these are equally the fault of the user and not the
platform, whereas things like rpc errors, panics, and other errors we aren't
expecting are 'our fault'. maybe my logic is sound, 'it beats nothing' maybe.

this adds a tag for blame=[service|user] which defaults to 'service' unless
set by invoke or http trigger, by the above standard. the plumbing in invoke
and http trigger sucks because we don't have the gin context thing to mutate
down in the invoke function we have (to make it easier to work with our fancy
response writers), so this is a bit blah, esp if we start whitelisting certain
errors, but it's a start.

this adds the concept of func errors, which are errors that can be blamed on a
function rather than the service, eg for things like a function hitting the
timeout due to being inside the container for too long to execute. to do this,
need to bubble up this field through grpc as these errors need to be
distinguished runner-side but consumed lb-side for metrics. this adds a new
type that builds on api errors, with the added distinction for function
specific errors, and transmits over grpc to rebuild the error's type on the lb
side.

adds func error for dispatch errors
make uds timeout a func error
make docker timeout/few more errors func errors

* make panicWrap last

it's inevitable we had to pay attention to this order eventually, here we are.

i tested this by putting a panic() in invoke and originally it did not appear
in metrics at all, now it does

* add log line with user error to call finish