WIP: permit the virtualised instantiation of Placers
This is part of an experiment to see if we can take real Fn code and
run it in a discrete event simulation with virtualised time.

Almost all of the naive placer (for instance) is amenable to this;
it leans on a placerTracker which does a hard sleep during backoff.

We can inject a virtualised version of this that runs off fake time.