Send RunnerMsg_ResultStart message
This change adds a call to a function to send the RunnerMsg_ResultStart
message during the enqueueCallResponse function.
The RunnerMsg_ResultStart contains any headers set by the function and
the status code.
This fixes the case where we don't send custom headers if a function
doesn't return a body.

Fixes #1413