make ch shed load under 503/error
previously we were consistently hashing to the same node even if we were full
on that node we were trying it with 100% probability and when it's loaded it
would with a high probability fail, this is a somewhat easy addition to shed
load with some probability when requests start failing against any given node,
using ewma, which will bounce up or down based on requests succeeding or
failing over time. when load goes down, probability will go up, otherwise it
will get shed across nodes.

this is not intended to be a very perfect solution and most of the values
themselves should be futzed with via simulation, however it should do better
than the naive ch algorithm if it works as intended, I haven't tested, this
was mostly for illustration and hopefully inspiration... one point of interest
is that this will loop back around on the ring and when all nodes are at
capacity this is likely to shed load somewhat evenly across nodes but it may
take a lot of retries, didn't optimize that case, but also not sure if we need
to as nodes should be scaling up when this happens?