fn: make evictor less aggressive
Experiments in load testing revealed that eviction of
containers that are starting is detrimental since this
increases churn and acts as a feedback mechanism
increasing docker API rates. Starting and deleting
a container are one of the most expensive docker API
calls and traffic arriving on a busy/loaded server
can trigger flood of container create actions that
get canceled with evictions. Previous code tried
to avoid this by tracking original request that
triggered the container start, but based on the
data this does not seem enough to avoid flurry
of evictions. In other words, in a busy system,
the likelihood of the original request getting
quickly serviced by another container is not
that rare.

With this change, we restrict evictability of
a container to Idle state exclusively. This makes
the backpressure (503) more likely since it
allows starting containers to initialize.

A race condition that occasionally causes busy
container eviction is also fixed. The fix proposed
here is that we unblock the listeners as if busy-container
really got evicted, but in busy container we simply
refresh the eviction token and get rid of the evicted
old token.

In future, based on empirical data, we may consider
introducing evictions for slow docker pulls.