Better management of gRPC Unavailable error code
It could happen that the LB looses the connection to a runner due to
some temporary network issue.
If that happens the client (loadbalancer) will get a gRPC Unavailable
error code back.
From https://godoc.org/google.golang.org/grpc/codes that error is
`likely a transient error and may be corrected by retrying`.
If we intercept one of those errors, instead of sending back a generic
500 (InternalServer error) we reply with a 502 (BadGateway).