Don't log stack trace at the top handler error if we have gRPG errors (#1424)
* Don't log stacktrace when we get gRPC errors

In the top error handler we log the stacktrace if we get back a non-API errors. Under that category we will have gRPC errors as well, there is no point to log a stacktrace for a gRPC errors where those could happen on the server side or due to some temporary network blips.
In this change we skip the stack trace log in presence of gRPC errors, plus we change some log level from Error down to Info for some retryable errors.