allow private registries to specify raw host
see the linked issue for a good example, it's kind of confusing. docker
defaults to not using a scheme, but the way we were preprocessing registry
urls meant that we were doing this incorrectly (it would parse out `u.Host = ''`
which is incorrect), even though we should allow this since it's dockers own
format and the registry parser itself allows this.

it seems like the reason we're pre-processing is for the subdomain logic,
which kind of makes sense, though it exhibits odd behaviors when specifying a
registry like `registry.com` when only 2 subdomains `raw.registry.com` and
`my.registry.com` have been registered, it's not deterministic between reboots
which credentials will be used, not sure how to reconcile this, I was unable
to find an issue pointing to why this came into existence with some cursory
searching, perhaps someone knows.

closes #1354