initialize ids in init (#1456)
this moves initialization up to the init() level as some packages were using
this to create ids before server.New() was called/finished (concurrency!), and
there was a race condition between setting this and creating ids.  this ought
to be sufficient while we're using the host ip address to seed these things,
that day may not last forever but at least this closes the hole for now.