add metrics to read data access thing
also fixes metrics, need to use the req context instead of the gin context in
order to get spans. more thumbs down for ginny.

this wraps the read data access on the whole, since the singleflight thing can
show up as latency around accessing the datastore, but they are not
necessarily the same thing (eg if 2 reqs come in, 1 waits for the other, but
won't have a latency for actually hitting the datastore, it's just waiting for
the cache to fill - instead of showing a void, show this to make it clear
where the request is).