make detached writer concurrency safe
in detached mode we're handing off execution but not interested in the headers
or body that the function writes at all. at present however, we're handing
over the original response headers for the detached writer to use, which could
happen concurrently when we're writing headers back to the client response.
since we're not interested in them, this just stops passing them from the
client response headers down into the agent for the func to write to, so there
should no longer be a race with the front end at least. the other piece was
that writing the status code in and reading it out is also not safe, exactly,
so made this explicitly safe - the case here would be that the function writes
a response code while we're going to write the response code in the front end.

I am not sure whether I like the way we're doing this or not, overall (prior
to the modifications here even). open to ideas.

updated the makefile/system test script so that I could run these faster to
repro, pretty handy, should add to other stuff too...