strip transport headers from function response
we weren't stripping transport headers from function response before copying
to the client response. this means e.g. a function could send a "Connection:
close" when fn wants to use a keep alive.

this also moved this to common since we're doing it in 3 places. I know that
it's not the ideal ideal big O performance and we can pedant about it, but
we're talking about 7 things here. I'm open to ideas about what this should
look like, I ripped it out of the stdlib and it seems okay. changed the
trigger stuff to use this too, as it was copied. added tests to cover all 3 of
these spots.

the only behavior change is no longer stripping 'authorization' header when
passing headers from the client request into the function. authorization seems
like a useful header for users that want to use auth in their functions, I
think if we need to use this one in the backend we should more carefully strip
out the parts we use before passing them on instead of relying on the agent to
strip the header to not leak anything into a function.

closes #1488